fix : 피드백 반영
