+++ src/main/java/com/kdt/bulletinboard/controller/MyExceptionHandler.java
@@ -1,0 +1,24 @@
+package com.kdt.bulletinboard.controller;
+
+import com.kdt.bulletinboard.ApiResponse;
+import javassist.NotFoundException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@Slf4j
+@ControllerAdvice
+public class MyExceptionHandler {
+
+    @ExceptionHandler
+    private ApiResponse<String> exceptionHandle(Exception exception) {
+        return ApiResponse.fail(HttpStatus.INTERNAL_SERVER_ERROR.value(), exception.getMessage());
+    }
+
+    @ExceptionHandler(NotFoundException.class)
+    private ApiResponse<String> notFoundHandle(NotFoundException exception) {
+        return ApiResponse.fail(HttpStatus.NOT_FOUND.value(), exception.getMessage());
+    }
+
+}
