+++ src/main/java/com/kdt/bulletinboard/dto/PostDto.java
@@ -5,6 +5,11 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
+
 @Getter
 @Builder
 @AllArgsConstructor
@@ -12,8 +17,15 @@ import lombok.NoArgsConstructor;
 public class PostDto {
 
     private Long id;
+
+    @NotBlank
+    @Size(min = 2)
     private String title;
+
+    @NotBlank
     private String content;
+
+    @NotBlank
     private UserDto userDto;
 
 }
