+++ src/main/java/com/kdt/bulletinboard/dto/UserDto.java
@@ -6,6 +6,10 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
+
 @Getter
 @Builder
 @AllArgsConstructor
@@ -13,7 +17,12 @@ import lombok.NoArgsConstructor;
 public class UserDto {
 
     private Long id;
+
+    @NotBlank
+    @Size(min = 2)
     private String name;
+
+    @NotBlank
     private Hobby hobby;
 
 }
