+++ src/main/java/com/kdt/bulletinboard/service/UserService.java
@@ -1,6 +1,5 @@
 package com.kdt.bulletinboard.service;
 
-import com.kdt.bulletinboard.converter.PostConverter;
 import com.kdt.bulletinboard.converter.UserConverter;
 import com.kdt.bulletinboard.dto.UserDto;
 import com.kdt.bulletinboard.entity.User;
@@ -10,17 +9,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import javax.transaction.Transactional;
 
 @Service
 public class UserService {
 
     @Autowired
     private UserRepository userRepository;
-
-    @Autowired
-    private PostConverter postConverter;
 
     @Autowired
     private UserConverter userConverter;
@@ -32,14 +28,14 @@ public class UserService {
         return saved.getId();
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public UserDto findOneUser(Long id) throws NotFoundException {
         return userRepository.findById(id)
                 .map(userConverter::convertToUserDto)
                 .orElseThrow(() -> new NotFoundException("해당 유저를 찾을 수 없습니다."));
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public Page<UserDto> findAllPost(Pageable pageable) {
         return userRepository.findAll(pageable)
                 .map(userConverter::convertToUserDto);
